backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
google-generativeai
requests
pydantic